capture: add memory alloc and free capture frames
This adds a new SysprofCaptureMemoryAlloc and
SysprofCaptureMemoryFree capture frame types. The Alloc frame also
allows setting the backtrace that led to the allocation in addition
to the thread identifier (such as from gettid).

The goal for this initially is to track fragmentation, but ultimately
we could use it for tracking allocations in general in terms of flame
graphs and similar.