memory: add preload to write out to sysprof capture
This is a very barebones implementation to do some capture recording
of memory allocations. It only uses a single writer with a lock which
appears to be fast enough for running a GTK app. However we may want
to do a multi-threaded solution with TLS later and multiple writers.

Still to do is the backtrace code to get instruction pointers for the
current stack.