capture: capture backtraces from libsysprof-capture
This allows us to capture the backtrace directly into the binary write
buffer instead of on the stack with a memcpy(). The side effect of that is
that we have to link against libunwind (if supported) from
libsysprof-capture.a. If libunwind is not available, we use backtrace()
from execinfo.h which is a bit slower it appears for the stacks I've
tested with (on Sysprof itself).

It would be nice if we could have a really fast unwinder even if it wasn't
very complete in handling DWARF/external debugging information.