memprof: use pixbuf drawing
We will have to draw a lot of these, so by reducing this to pixel drawing
we can save quite a bit of overhead and memory from storing paths for
eventual filling.

Also, we should have enough pixels to draw that you'll still get the idea
of allocations without having to draw lines.